ROFO-112 refactor: TokenState model 패키지로 이동
