+++ domain/src/main/java/com/weit2nd/domain/model/token/TokenState.kt
@@ -1,4 +1,4 @@
-package com.weit2nd.domain.exception.token
+package com.weit2nd.domain.model.token
 
 sealed class TokenState {
     data object AccessTokenValid : TokenState()
