+++ domain/src/main/java/com/weit2nd/domain/repository/login/LoginRepository.kt
@@ -1,6 +1,6 @@
 package com.weit2nd.domain.repository.login
 
-import com.weit2nd.domain.exception.token.TokenState
+import com.weit2nd.domain.model.token.TokenState
 
 interface LoginRepository {
     suspend fun loginWithKakao(): Result<Unit>
