+++ domain/src/main/java/com/weit2nd/domain/usecase/login/GetTokenStateUseCase.kt
@@ -1,6 +1,6 @@
 package com.weit2nd.domain.usecase.login
 
-import com.weit2nd.domain.exception.token.TokenState
+import com.weit2nd.domain.model.token.TokenState
 import com.weit2nd.domain.repository.login.LoginRepository
 import javax.inject.Inject
